+++ src/apps/seelenweg/modules/shared/store/domain.ts
@@ -1,5 +1,4 @@
 import { SeelenWegSettings, WegItem } from '@seelen-ui/lib/types';
-import { modify } from 'readable-types';
 
 import { IRootState } from '../../../../../shared.interfaces';
 import { FocusedApp } from '../../../../shared/interfaces/common';
@@ -37,31 +36,13 @@ export interface MediaSession {
 }
 
 export type PinnedWegItem = Extract<WegItem, { type: 'Pinned' }>;
-export type ExtendedPinnedWegItem = modify<
-  PinnedWegItem,
-  {
-    icon: string;
-    title: string;
-    opens: HWND[];
-  }
->;
-
-export type ExtendedTemporalWegItem = modify<
-  ExtendedPinnedWegItem,
-  {
-    type: 'Temporal';
-  }
->;
-
+export type TemporalWegItem = Extract<WegItem, { type: 'Temporal' }>;
 export type SeparatorWegItem = Extract<WegItem, { type: 'Separator' }>;
 export type MediaWegItem = Extract<WegItem, { type: 'Media' }>;
 export type StartMenuWegItem = Extract<WegItem, { type: 'StartMenu' }>;
-export type SwItem =
-  | ExtendedPinnedWegItem
-  | ExtendedTemporalWegItem
-  | SeparatorWegItem
-  | MediaWegItem
-  | StartMenuWegItem;
+
+/** @alias */
+export type SwItem = WegItem;
 
 export interface RootState extends IRootState<SeelenWegSettings> {
   devTools: boolean;